refactor: move aliases in WebHandler
